software-testing-2025-advanced/requirements.txt
pytest==8.3.2
pytest-html==4.1.1
pytest-cov==5.0.0
requests==2.32.3
jsonschema==4.23.0
hypothesis==6.112.1
PyYAML==6.0.2
selenium==4.23.1
webdriver-manager==4.0.2
